backend/requirements.txt
Django==3.2.14
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0
environs==9.5.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-quotes==3.3.1
flake8-return==1.1.3
gunicorn==20.1.0
isort==5.10.1
pep8-naming==0.13.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytest==7.1.2
reportlab==3.6.2